Handle graph weirdness with git bisect visualize
We use git bisect visualize to find all the commits that bisect thinks it can visit. Then we use that as our "unknown" shas. We then can flag commits as good/bad from this state using a 2 tree walking methods, one that finds all branches to a target sha, and one that finds all reachable shas from a given commit.